feat: build MDX slides with next
Use MDX-Deck's babel config and webpack setup. This is difficult
due to mdx-deck's internals which rely on separate router (Reach).
This will render the page but break on refresh on any page outside
of the index. This is because NextJS needs some help to manage its
path parameters, usually in the form of extending the express server.